feat: KC integration within endpoints

+++ src/main/java/com/tidsbankencaseapi/Controllers/EmployeeController.java
@@ -4,45 +4,62 @@ import com.tidsbankencaseapi.Models.Employee;
 import com.tidsbankencaseapi.Models.VacationRequest;
 import com.tidsbankencaseapi.Repositories.EmployeeRepository;
 import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.web.bind.annotation.*;
 
+import java.net.URI;
+import java.net.URISyntaxException;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.UUID;
 
 @RestController
 @RequestMapping("/employee")
+@SecurityRequirement(name = "keycloak_implicit")
 public class EmployeeController {
 
     @Autowired
     private EmployeeRepository employeeRepository;
 
     //GET /employee
-    /*@Operation(summary = "Get Employee")
+    @Operation(summary = "Get Employee")
     @GetMapping("")
-    public ResponseEntity<Employee> getEmployee() {
-        HttpStatus status;
+    @PreAuthorize("hasAnyRole('user', 'administrator')")//ADMIN / USER protected
+    public ResponseEntity<String> getEmployee(@AuthenticationPrincipal Jwt principal) throws URISyntaxException {
 
-        //AUTH
-        //get logged in employee / Can we get their id as well?
-        //Integer id = keycloak get employee id
-        if (!employeeRepository.existsById(id) {
+        HttpStatus status;
+        String currentEmployeeId = principal.getClaimAsString("sub");
+
+        URI location = new URI(String.format("https://api-tidsbanken-case.herokuapp.com/employee", currentEmployeeId));
+        HttpHeaders headers = new HttpHeaders();
+
+        if (!employeeRepository.existsById(currentEmployeeId)) {
             status = HttpStatus.BAD_REQUEST;
         } else {
             status = HttpStatus.SEE_OTHER;
-
+            headers.setLocation(location);
+            headers.setBearerAuth(String.valueOf(principal));
         }
-    }*/
+        return new ResponseEntity<>(headers, status);
+    }
 
     //POST /employee
     @Operation(summary = "Register Employee")
     @PostMapping("/register")
+    @PreAuthorize("hasRole('administrator')") //ADMIN protected
     public ResponseEntity<Employee> registerEmployee(@RequestBody Employee employee) {
+
         HttpStatus status;
-        //AUTH
-        //Put employee in keycloak
+
+        //always init with admin false
+        employee.isAdmin = Boolean.FALSE;
 
         if (!employeeRepository.existsById(employee.employeeId)) {
             employee = employeeRepository.save(employee);
@@ -56,17 +73,25 @@ public class EmployeeController {
     //GET /employee/:employee_id
     @Operation(summary = "Get Employee by ID")
     @GetMapping("/{id}")
-    public ResponseEntity<Employee> getEmployeeId(@PathVariable Integer id) {
-        Employee employee = new Employee();
+    @PreAuthorize("hasAnyRole('user', 'administrator')")//ADMIN / USER protected
+    public ResponseEntity<Employee> getEmployeeId(@PathVariable String id, @AuthenticationPrincipal Jwt principal) {
+
         HttpStatus status;
-        //AUTH
-        //if user should return only pic and name
+        Employee employee = new Employee();
+        List<String> signedInRole = principal.getClaimAsStringList("roles");
 
         if (!employeeRepository.existsById(id)) {
             status = HttpStatus.NOT_FOUND;
         } else {
-            status = HttpStatus.OK;
-            employee = employeeRepository.findById(id).get();
+            if (signedInRole.contains("administrator")) {
+                employee = employeeRepository.findById(id).get();
+                status = HttpStatus.OK;
+            } else {
+                employee.first_name = employeeRepository.findById(id).get().first_name;
+                employee.last_name = employeeRepository.findById(id).get().last_name;
+                employee.profilePic = employeeRepository.findById(id).get().profilePic;
+                status =HttpStatus.OK;
+            }
         }
 
         return new ResponseEntity<>(employee, status);
@@ -75,18 +100,47 @@ public class EmployeeController {
     //PATCH /employee/:employee_id
     @Operation(summary = "Update Employee")
     @PatchMapping("/update/{id}")
-    public ResponseEntity<Employee> updateEmployee(@PathVariable Integer id, @RequestBody Employee employee) {
-        //AUTH
-        //update of password returns bad request
-        //only admin can change isAdmin all other unauth attempts respond with 403 Forbidden
-        Employee returnEmployee = new Employee();
-        HttpStatus status;
+    @PreAuthorize("hasAnyRole('user', 'administrator')")//ADMIN / USER protected
+    public ResponseEntity<Employee> updateEmployee(@PathVariable String id, @RequestBody Employee employee, @AuthenticationPrincipal Jwt principal) {
 
+        HttpStatus status;
+        Employee returnEmployee = employeeRepository.findById(id).get();
+        List<String> signedInRole = principal.getClaimAsStringList("roles");
+
+        //check if pathvariable id equals request employee id
         if (!id.equals(employee.employeeId)) {
             status = HttpStatus.BAD_REQUEST;
             return new ResponseEntity<>(returnEmployee, status);
+        } else {
+
+            //if admin is changed by admin change it else return 403 FORBIDDEN.
+            if (employee.isAdmin != null)  {
+                if (signedInRole.contains("administrator")) {
+                    returnEmployee.isAdmin = employee.isAdmin;
+                } else  {
+                    status =  HttpStatus.FORBIDDEN;
+                    return new ResponseEntity<>(returnEmployee, status);
+                }
+            }
+
+            if (employee.first_name != null) {
+                returnEmployee.first_name = employee.first_name;
+            }
+
+            if (employee.last_name != null) {
+                returnEmployee.last_name = employee.last_name;
+            }
+
+            if (employee.emailAddress != null) {
+                returnEmployee.emailAddress = employee.emailAddress;
+            }
+
+            if (employee.profilePic != null) {
+                returnEmployee.profilePic = employee.profilePic;
+            }
         }
-        returnEmployee = employeeRepository.save(employee);
+
+        employeeRepository.save(returnEmployee);
         status = HttpStatus.NO_CONTENT;
         return new ResponseEntity<>(returnEmployee, status);
     }
@@ -94,16 +148,24 @@ public class EmployeeController {
     //DELETE /employee/:employee_id
     @Operation(summary = "Delete an Employee")
     @DeleteMapping("/delete/{id}")
-    public ResponseEntity<Employee> deleteEmployee(@PathVariable Integer id) {
-        HttpStatus status;
-        //AUTH
-        //only self or admin can delete
+    @PreAuthorize("hasAnyRole('user', 'administrator')")//ADMIN / USER protected
+    public ResponseEntity<Employee> deleteEmployee(@PathVariable String id, @AuthenticationPrincipal Jwt principal) {
 
+        HttpStatus status;
+        String currentEmployeeId = principal.getClaimAsString("sub");
+        List<String> signedInRole = principal.getClaimAsStringList("roles");
+
+        //if employee does not exist
         if (!employeeRepository.existsById(id)) {
             status = HttpStatus.BAD_REQUEST;
         } else {
-            employeeRepository.deleteById(id);
-            status = HttpStatus.OK;
+            //if user asking is admin or self delete else 403-FORBIDDEN
+            if (signedInRole.contains("administrator") || currentEmployeeId == id) {
+                employeeRepository.deleteById(id);
+                status = HttpStatus.OK;
+            } else {
+                status = HttpStatus.FORBIDDEN;
+            }
         }
         return new ResponseEntity<>(status);
     }
@@ -111,7 +173,8 @@ public class EmployeeController {
     //GET /employee/:employee_id/requests
     @Operation(summary = "Get an specific Employee's requests")
     @GetMapping("/{id}/requests")
-    public ResponseEntity<List<VacationRequest>> getEmployeeRequests(@PathVariable Integer id) {
+    @PreAuthorize("hasAnyRole('user', 'administrator')")//ADMIN / USER protected
+    public ResponseEntity<List<VacationRequest>> getEmployeeRequests(@PathVariable String id, @AuthenticationPrincipal Jwt principal) {
         //Optionally accepts appropriate query parameters to search and limit the number of objects returned.
 
         HttpStatus status;
@@ -120,19 +183,9 @@ public class EmployeeController {
         if (!employeeRepository.existsById(id)) {
             status = HttpStatus.BAD_REQUEST;
         } else  {
-            Employee employee = employeeRepository.findById(id).get();
-            listRequests = employee.getVacationRequests();
+            listRequests = employeeRepository.findById(id).get().getVacationRequests();
             status = HttpStatus.OK;
         }
         return new ResponseEntity<>(listRequests, status);
     }
-
-    //POST /employee/:employee_id/update_password
-    /*@Operation(summary = "Change Password")
-    @PostMapping("/{id}/update_password")
-    public ResponseEntity<Employee> updatePassword(@PathVariable Integer id, @RequestBody String Password) {
-        //AUTH
-        //Send new password to keycloak
-    }*/
-
 }
